Broke down FalcorRouter into folder structure
