+++ src/lib/falcor/routes/articles/index.js
@@ -1,0 +1,7 @@
+import articlesByPage from './articles-by-page';
+import articlesBySlug from './articles-by-slug';
+
+export default [].concat(
+  articlesByPage,
+  articlesBySlug
+);
