+++ src/lib/falcor/routes/authors/index.js
@@ -1,0 +1,3 @@
+import authorsBySlug from './authors-by-slug';
+
+export default authorsBySlug;
