+++ src/lib/falcor/routes/categories/index.js
@@ -1,0 +1,7 @@
+import categoriesByIndex from './categories-by-index';
+import categoriesBySlug from './categories-by-slug';
+
+export default [].concat(
+  categoriesByIndex,
+  categoriesBySlug
+);
