+++ src/lib/falcor/routes/index.js
@@ -1,0 +1,25 @@
+import articles from './articles';
+import authors from './authors';
+import categories from './categories';
+import infoPages from './info-pages';
+import issues from './issues';
+import placeholder from './placeholder';
+import search from './search';
+import teams from './teams';
+import trending from './trending';
+import totalAmountOfArticles from './total-amount-of-articles';
+import latestIssue from './latest-issue';
+
+export {
+  articles,
+  authors,
+  categories,
+  infoPages,
+  issues,
+  placeholder,
+  search,
+  teams,
+  trending,
+  totalAmountOfArticles,
+  latestIssue,
+};
