+++ src/lib/falcor/routes/issues/index.js
@@ -1,0 +1,3 @@
+import issuesByNumber from './issues-by-number';
+
+export default issuesByNumber;
