+++ src/lib/falcor/routes/placeholder.js
@@ -1,0 +1,12 @@
+export default [
+  {
+    // A placeholder so we can do empty calls that just invalidate
+    route: 'placeholder',
+    get: () => (
+      [{
+        path: ['placeholder'],
+        value: 'placeholder',
+      }]
+    ),
+  },
+];
