+++ src/lib/falcor/routes/search/authors.js
@@ -1,0 +1,47 @@
+import falcor from 'falcor';
+import _ from 'lodash';
+
+import DbFunctions from 'lib/db';
+
+const db = new DbFunctions;
+const $ref = falcor.Model.ref;
+
+export default [
+  {
+    // Search for authors by name
+    route: "search['authors'][{keys:queries}][{integers:indices}]",
+    get: (pathSet) => (
+      new Promise((resolve) => {
+        let minIndex = pathSet.indices[0];
+        let maxIndex = pathSet.indices[0];
+        pathSet.indices.forEach((index) => {
+          if (index < minIndex) {
+            minIndex = index;
+          }
+          if (index > maxIndex) {
+            maxIndex = index;
+          }
+        });
+        db.searchAuthorsQuery(pathSet.queries, minIndex, maxIndex).then((data) => {
+          // Map all the indices down to fit the indices returned by the db call
+          const processedPathSet = {
+            ...pathSet,
+            indices: pathSet.indices.map(index => index - minIndex),
+          };
+          const results = [];
+          _.forEach(data, (queryResults, query) => {
+            processedPathSet.indices.forEach((index) => {
+              if (index < queryResults.length) {
+                results.push({
+                  path: ['search', 'authors', query, index],
+                  value: $ref(['authorsBySlug', queryResults[index]]),
+                });
+              }
+            });
+          });
+          resolve(results);
+        });
+      })
+    ),
+  },
+];
