+++ src/lib/falcor/routes/search/index.js
@@ -1,0 +1,9 @@
+import authors from './authors';
+import posts from './posts';
+import teams from './teams';
+
+export default [].concat(
+  authors,
+  posts,
+  teams
+);
