+++ src/lib/falcor/routes/teams/index.js
@@ -1,0 +1,7 @@
+import teamsBySlug from './teams-by-slug';
+import teamsByIndex from './teams-by-index';
+
+export default [].concat(
+  teamsByIndex,
+  teamsBySlug
+);
